feat[#73] : 유저 세션 관리 로직 모두 클라이언트단으로 이동
