+++ blog/app/temp/[articleId]/page.tsx
@@ -3,24 +3,9 @@ import React from "react";
 
 import { getArticleById } from "@/entities/article/model";
 
-import { createBrowserSupabase } from "@/shared/model";
-
 interface TempArticlePageProps {
   params: Promise<{ articleId: string }>;
 }
-
-export async function generateStaticParams() {
-  const supabase = createBrowserSupabase();
-
-  const { data: ids } = await supabase
-    .from("articles")
-    .select("id")
-    .eq("status", "published");
-
-  return ids ? ids.map(({ id }) => ({ id })) : [];
-}
-export const dynamic = "force-static";
-export const dynamicParams = false;
 
 const TempArticlePage: React.FC<TempArticlePageProps> = async ({ params }) => {
   const { articleId } = await params;
