+++ blog/app/temp/layout.tsx
@@ -1,0 +1,25 @@
+"use client";
+
+import { useRouter } from "next/navigation";
+import { PropsWithChildren, useEffect } from "react";
+
+import { useSession } from "@/entities/user/model";
+
+const TempLayout: React.FC<PropsWithChildren> = ({ children }) => {
+  const { user } = useSession();
+  const router = useRouter();
+
+  useEffect(() => {
+    if (!user) {
+      router.push("/auth");
+    }
+  }, [user, router]);
+
+  if (!user) {
+    return null;
+  }
+
+  return children;
+};
+
+export default TempLayout;
