--- blog/src/entities/user/model/getAuthorizedUser.ts
@@ -1,6 +1,0 @@
-import { createServerSupabase } from "@/shared/model";
-
-export const getAuthorizedUser = async () => {
-  const supabase = await createServerSupabase();
-  return supabase.auth.getUser();
-};
