--- blog/src/features/auth/lib/index.ts
@@ -1,2 +1,0 @@
-export * from "./logoutAction";
-export * from "./useLoginAction";
