+++ blog/src/shared/route/index.ts
@@ -1,4 +1,3 @@
-export * from "./supabaseMiddleware";
 export * from "./camelToSnake";
 export * from "./createPostgressErrorResponse";
 export * from "./createImageConfig";
